Remove calendar caching
This was leading to the calendar data being consistenlty read from the
cache instead of the actual ics file in all but the first time loading
the file.
As a result, the date limits were not taken into consideration as only
filtered data was dumped to the cache and it wasn't re-filtered on
loading.